Meta(dotenv & directory): Update absolute NODE_PATH and component import structure
