+++ my-react/.env
@@ -1,1 +1,4 @@
 SKIP_PREFLIGHT_CHECK=true
+
+# Set Absolute path for WebPack Path Resolver
+NODE_PATH=src/
