+++ my-react/src/componentBlocks/index.js
@@ -1,15 +1,1 @@
-import StyledMain from './StyledTodo/Main';
-import StyledHeader from './StyledTodo/Header';
-import StyledTodoInput from './StyledTodo/TodoInput';
-import StyledTodoItem from './StyledTodo/TodoItem';
-import StyledTodoList from './StyledTodo/FoldableList/TodoList';
-
-const StyledTodo = {
-  Main: StyledMain,
-  Header: StyledHeader,
-  Input: StyledTodoInput,
-  Item: StyledTodoItem,
-  FoldableList: StyledTodoList,
-};
-
-export default StyledTodo;
+export { default as StyledTodo } from './StyledTodo';
